Home

FresnelIntegrals.jl

FresnelIntegrals.fresnel

FresnelIntegrals.fresnelc

FresnelIntegrals.fresnels

FresnelIntegrals.fresnel — Method.

    fresnel(z)

Calculates the cosine and sine fresnel integrals

source

FresnelIntegrals.fresnelc — Method.

    fresnelc(z)

Calculates the Fresnel cosine integral for the number z

source

FresnelIntegrals.fresnels — Method.

    fresnels(z)

Calculates the Fresnel sine integral for the number z

source