Home

FresnelIntegrals.jl

A quick Julia library for calculating Fresnel Integrals using the error function from SpecialFunctions.jl.

Installation

Until this is added into Pkg's manifest, just add from the Github link

Pkg.add("https://github.com/kiranshila/FresnelIntegrals.jl")

Functions

FresnelIntegrals.fresnel — Method.

    fresnel(z)

Calculates the cosine and sine fresnel integrals

source

FresnelIntegrals.fresnelc — Method.

    fresnelc(z)

Calculates the Fresnel cosine integral for the number z for $C(z) = \int_{0}^{\infty} \cos{\left(\frac{\pi t^2}{2}\right)}$

source

FresnelIntegrals.fresnels — Method.

    fresnels(z)

Calculates the Fresnel sine integral for the number z for $S(z) = \int_{0}^{\infty} \sin{\left(\frac{\pi t^2}{2}\right)}$

source